SiMBA Implementation

+++ ferret/mbablastprovider.py
@@ -164,8 +164,8 @@ class MBABlastEqualityProvider(EqualityProvider):
     
 
     def _truthtable_to_coefficients(self, truthtable, var_amount):
-        A = np.asmatrix(self.truthbasisMap[var_amount]).T
-        b = np.asmatrix(truthtable).T
+        A = np.asmatrix(self.truthbasisMap[var_amount], dtype=np.int64).T
+        b = np.asmatrix(truthtable, dtype=np.int64).T
         r = np.linalg.solve(A, b)
         l = [round(i) for i in np.array(r).reshape(-1,).tolist()]
         # would be smart to cache l for truthtable key
